pyproject.toml
[project]
name = "wallet-service-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.0",
    "passlib[bcrypt]>=1.7.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.5.3",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.8",
]
